+++ rocket-okapi-codegen/src/openapi_attr/mod.rs
@@ -9,35 +9,50 @@ use quote::quote;
 use quote::ToTokens;
 use rocket_http::Method;
 use std::collections::BTreeMap as Map;
-use syn::{parse_macro_input, AttributeArgs, FnArg, Ident, ItemFn, ReturnType, Type, TypeTuple};
+use syn::{
+    parse_macro_input, AttributeArgs, FnArg, GenericArgument, Ident, ItemFn, PathArguments,
+    PathSegment, ReturnType, Type, TypeTuple,
+};
 
 #[derive(Debug, Default, FromMeta)]
 #[darling(default)]
 struct OpenApiAttribute {
+    /// Skip this API entry and do not add it to the documentation.
     pub skip: bool,
 
+    /// Add tags to the entry to group them together.
     #[darling(multiple, rename = "tag")]
     pub tags: Vec<String>,
-    pub id: Option<String>,
+  
+    /// `operationId` is an optional unique string used to identify an operation.
+    /// If provided, these IDs must be unique among all operations described in your API.
+    pub operation_id: Option<String>,
+
+    /// Ignore certain parameters from the function and do not include them in the documentation.
+    #[darling(multiple)]
+    pub ignore: Vec<String>,
+
+    /// Mark this operation as deprecated in the documentation.
+    pub deprecated: bool,
 }
 
 pub fn parse(args: TokenStream, input: TokenStream) -> TokenStream {
     let attr_args = parse_macro_input!(args as AttributeArgs);
     let input = parse_macro_input!(input as ItemFn);
 
-    let okapi_attr = match OpenApiAttribute::from_list(&attr_args) {
+    let entry_attributes = match OpenApiAttribute::from_list(&attr_args) {
         Ok(v) => v,
         Err(e) => {
             return e.write_errors().into();
         }
     };
 
-    if okapi_attr.skip {
+    if entry_attributes.skip {
         return create_empty_route_operation_fn(input);
     }
 
     match route_attr::parse_attrs(&input.attrs) {
-        Ok(route) => create_route_operation_fn(input, route, okapi_attr),
+        Ok(route) => create_route_operation_fn(input, route, &entry_attributes),
         Err(e) => e,
     }
 }
@@ -45,6 +60,7 @@ pub fn parse(args: TokenStream, input: TokenStream) -> TokenStream {
 fn create_empty_route_operation_fn(route_fn: ItemFn) -> TokenStream {
     let fn_name = get_add_operation_fn_name(&route_fn.sig.ident);
     TokenStream::from(quote! {
+        #[doc(hidden)]
         pub fn #fn_name(
             _gen: &mut ::rocket_okapi::gen::OpenApiGenerator,
             _op_id: String,
@@ -54,14 +70,64 @@ fn create_empty_route_operation_fn(route_fn: ItemFn) -> TokenStream {
     })
 }
 
+/// Replace `EventStream<impl SOMETHING>`
+/// with `EventStream`
+fn type_replace_impl_trait(ty: Type) -> Type {
+    if let Type::Path(type_path) = &ty {
+        if let Some(path_segment) = type_path.path.segments.first() {
+            if let PathArguments::AngleBracketed(generic_argument) = &path_segment.arguments {
+                if let Some(generic_argument) = generic_argument.args.first() {
+                    if let Some(result_type) =
+                        _type_replace_impl_trait_generic_argument(generic_argument, path_segment)
+                    {
+                        return result_type;
+                    }
+                }
+            }
+        }
+    }
+    ty
+}
+
+/// Helper function for `type_replace_impl_trait` and should not be called from anywhere else.
+fn _type_replace_impl_trait_generic_argument(
+    gen_arg: &GenericArgument,
+    path_segment: &PathSegment,
+) -> Option<Type> {
+    if let GenericArgument::Type(Type::ImplTrait(_)) = gen_arg {
+        if path_segment.ident == "EventStream" {
+            // Return special type, the type of stream does not matter as long as something is present
+            return Some(Type::Verbatim(quote! {
+                EventStream<rocket::futures::stream::Empty<rocket::response::stream::Event>>
+            }));
+        } else if path_segment.ident == "ByteStream" {
+            // Return special type, the type of stream does not matter as long as something is present
+            return Some(Type::Verbatim(quote! {
+                ByteStream<rocket::futures::stream::Empty<Vec<u8>>>
+            }));
+        } else if path_segment.ident == "ReaderStream" {
+            // Return special type, the type of stream does not matter as long as something is present
+            return Some(Type::Verbatim(quote! {
+                ReaderStream<rocket::futures::stream::Empty<File>>
+            }));
+        } else if path_segment.ident == "TextStream" {
+            // Return special type, the type of stream does not matter as long as something is present
+            return Some(Type::Verbatim(quote! {
+                TextStream<rocket::futures::stream::Empty<String>>
+            }));
+        }
+    }
+    None
+}
+
 fn create_route_operation_fn(
     route_fn: ItemFn,
     route: route_attr::Route,
-    attr: OpenApiAttribute,
+    entry_attributes: &OpenApiAttribute,
 ) -> TokenStream {
     let arg_types = get_arg_types(route_fn.sig.inputs.into_iter());
     let return_type = match route_fn.sig.output {
-        ReturnType::Type(_, ty) => *ty,
+        ReturnType::Type(_, ty) => type_replace_impl_trait(*ty),
         ReturnType::Default => unit_type(),
     };
 
@@ -76,12 +142,26 @@ fn create_route_operation_fn(
     // Create a list of all the already used parameters.
     // This is later used to see what parameters are Request Guards. (aka left over)
     let mut params_names_used = Vec::new();
+    // Keep track of ignored arguments so we can give warnings in case the argument does not do anything.
+    let mut ignored_names = Vec::new();
     // Path parameters: `/<id>/<name>`
     for arg in route.path_params() {
+        // Do NOT skip all attributes that are in ignore list, because in this case it will make
+        // the API documentation incorrect.
+        if entry_attributes.ignore.contains(&arg.to_owned()) {
+            // Create compile error
+            return TokenStream::from(quote! {
+                compile_error!(concat!(
+                    "The argument `",
+                    #arg,
+                    "` can not be ignored, because it is a path parameter."
+                ));
+            });
+        }
         let ty = match arg_types.get(arg) {
             Some(ty) => ty,
             None => return quote! {
-                compile_error!(concat!("Could not find argument ", #arg, " matching path param."));
+                compile_error!(concat!("Could not find argument `", #arg, "` matching path param."));
             }
             .into(),
         };
@@ -92,24 +172,34 @@ fn create_route_operation_fn(
     }
     // Multi Path parameters: `/<path..>`
     if let Some(arg) = route.path_multi_param() {
-        let ty = match arg_types.get(arg) {
-            Some(ty) => ty,
-            None => return quote! {
-                compile_error!(concat!("Could not find argument ", #arg, " matching multi path param."));
-            }
-            .into(),
-        };
-        params_names_used.push(arg.to_owned());
-        params.push(quote! {
-            <#ty as ::rocket_okapi::request::OpenApiFromSegments>::path_multi_parameter(gen, #arg.to_owned())?.into()
-        })
+        // Skip all attributes that are in ignore list.
+        if entry_attributes.ignore.contains(&arg.to_owned()) {
+            ignored_names.push(arg.to_owned());
+        } else {
+            let ty = match arg_types.get(arg) {
+                Some(ty) => ty,
+                None => return quote! {
+                    compile_error!(concat!("Could not find argument `", #arg, "` matching multi path param."));
+                }
+                .into(),
+            };
+            params_names_used.push(arg.to_owned());
+            params.push(quote! {
+                <#ty as ::rocket_okapi::request::OpenApiFromSegments>::path_multi_parameter(gen, #arg.to_owned())?.into()
+            })
+        }
     }
     // Query parameters: `/?<id>&<name>`
     for arg in route.query_params() {
+        // Skip all attributes that are in ignore list.
+        if entry_attributes.ignore.contains(&arg.to_owned()) {
+            ignored_names.push(arg.to_owned());
+            continue;
+        }
         let ty = match arg_types.get(arg) {
             Some(ty) => ty,
             None => return quote! {
-                compile_error!(concat!("Could not find argument ", #arg, " matching query param."));
+                compile_error!(concat!("Could not find argument `", #arg, "` matching query param."));
             }
             .into(),
         };
@@ -120,10 +210,15 @@ fn create_route_operation_fn(
     }
     // Multi Query parameters: `/?<param..>`
     for arg in route.query_multi_params() {
+        // Skip all attributes that are in ignore list.
+        if entry_attributes.ignore.contains(&arg.to_owned()) {
+            ignored_names.push(arg.to_owned());
+            continue;
+        }
         let ty = match arg_types.get(arg) {
             Some(ty) => ty,
             None => return quote! {
-                compile_error!(concat!("Could not find argument ", #arg, " matching multi query param."));
+                compile_error!(concat!("Could not find argument `", #arg, "` matching multi query param."));
             }.into(),
         };
         params_names_used.push(arg.to_owned());
@@ -155,6 +250,11 @@ fn create_route_operation_fn(
     // https://rocket.rs/v0.5-rc/guide/requests/#request-guards
     // Request Guards is every that is not already used and thus not in `params_names_used`.
     for (arg, ty) in &arg_types {
+        // Skip all attributes that are in ignore list.
+        if entry_attributes.ignore.contains(arg) {
+            ignored_names.push(arg.to_owned());
+            continue;
+        }
         if !params_names_used.contains(arg) {
             params_names_used.push(arg.to_owned());
             params_request_guards.push(quote! {
@@ -166,14 +266,28 @@ fn create_route_operation_fn(
         }
     }
 
+    // Check if everything in ignore list is actually found.
+    for ignore_arg in &entry_attributes.ignore {
+        if !ignored_names.contains(ignore_arg) {
+            // Create compile error
+            return TokenStream::from(quote! {
+                compile_error!(concat!(
+                    "The argument `",
+                    #ignore_arg,
+                    "` does not exist and thus can not be ignored."
+                ));
+            });
+        }
+    }
+
     let fn_name = get_add_operation_fn_name(&route_fn.sig.ident);
     let path = route
         .origin
         .path()
         .as_str()
-        .replace("<", "{")
+        .replace('<', "{")
         .replace("..>", "}")
-        .replace(">", "}");
+        .replace('>', "}");
     let method = Ident::new(&to_pascal_case_string(route.method), Span::call_site());
     let (title, desc) = doc_attr::get_title_and_desc_from_doc(&route_fn.attrs);
     let title = match title {
@@ -185,20 +299,24 @@ fn create_route_operation_fn(
         None => quote!(None),
     };
 
-    let tags = attr
+    let tags = entry_attributes
         .tags
-        .into_iter()
+        .iter()
         .map(|tag| quote!(#tag.to_owned()))
         .collect::<Vec<_>>();
-    let op_id = match attr.id {
-        Some(id) => quote! { #id.into() },
-        None => quote! { op_id },
+  
+    let deprecated = entry_attributes.deprecated;
+  
+    let operation_id = match attr.operation_id {
+        Some(operation_id) => quote! { #operation_id.into() },
+        None => quote! { operation_id },
     };
 
     TokenStream::from(quote! {
+        #[doc(hidden)]
         pub fn #fn_name(
             gen: &mut ::rocket_okapi::gen::OpenApiGenerator,
-            op_id: String,
+            operation_id: String,
         ) -> ::rocket_okapi::Result<()> {
             let mut responses = <#return_type as ::rocket_okapi::response::OpenApiResponder>::responses(gen)?;
             // Add responses from Request Guards.
@@ -256,7 +374,7 @@ fn create_route_operation_fn(
                 path: #path.to_owned(),
                 method: ::rocket::http::Method::#method,
                 operation: ::rocket_okapi::okapi::openapi3::Operation {
-                    operation_id: Some(#op_id),
+                    operation_id: Some(#operation_id),
                     responses,
                     request_body,
                     parameters,
@@ -264,6 +382,7 @@ fn create_route_operation_fn(
                     description: #desc,
                     security,
                     tags: vec![#(#tags),*],
+                    deprecated: #deprecated,
                     ..Default::default()
                 },
             });