+++ rocket-okapi/src/request/from_data_impls.rs
@@ -33,11 +33,11 @@ macro_rules! fn_request_body {
     }};
 }
 
-impl<'r> OpenApiFromData<'r> for String {
-    fn request_body(gen: &mut OpenApiGenerator) -> Result {
-        fn_request_body!(gen, String, "application/octet-stream")
-    }
-}
+// Implement `OpenApiFromData` for everything that implements `FromData`
+// Order is same as on:
+// https://docs.rs/rocket/0.5.0-rc.2/rocket/data/trait.FromData.html#foreign-impls
+
+// ## Implementations on Foreign Types
 
 impl<'r> OpenApiFromData<'r> for &'r str {
     fn request_body(gen: &mut OpenApiGenerator) -> Result {
@@ -47,13 +47,7 @@ impl<'r> OpenApiFromData<'r> for &'r str {
 
 impl<'r> OpenApiFromData<'r> for Cow<'r, str> {
     fn request_body(gen: &mut OpenApiGenerator) -> Result {
-        fn_request_body!(gen, str, "application/octet-stream")
-    }
-}
-
-impl<'r> OpenApiFromData<'r> for Vec<u8> {
-    fn request_body(gen: &mut OpenApiGenerator) -> Result {
-        fn_request_body!(gen, Vec<u8>, "application/octet-stream")
+        <&'r str>::request_body(gen)
     }
 }
 
@@ -63,20 +57,7 @@ impl<'r> OpenApiFromData<'r> for &'r [u8] {
     }
 }
 
-impl<'r, T: OpenApiFromData<'r> + 'r> OpenApiFromData<'r> for StdResult<T, T::Error> {
-    fn request_body(gen: &mut OpenApiGenerator) -> Result {
-        T::request_body(gen)
-    }
-}
-
-impl<'r, T: OpenApiFromData<'r>> OpenApiFromData<'r> for Option<T> {
-    fn request_body(gen: &mut OpenApiGenerator) -> Result {
-        Ok(RequestBody {
-            required: false,
-            ..T::request_body(gen)?
-        })
-    }
-}
+// ## Implementors
 
 // Waiting for https://github.com/GREsau/schemars/issues/103
 impl<'r> OpenApiFromData<'r> for rocket::fs::TempFile<'r> {
@@ -84,40 +65,16 @@ impl<'r> OpenApiFromData<'r> for rocket::fs::TempFile<'r> {
         Vec::<u8>::request_body(gen)
     }
 }
+
 impl<'r> OpenApiFromData<'r> for rocket::data::Capped<rocket::fs::TempFile<'r>> {
     fn request_body(gen: &mut OpenApiGenerator) -> Result {
         rocket::fs::TempFile::request_body(gen)
     }
 }
+
 impl<'r> OpenApiFromData<'r> for rocket::data::Capped<Cow<'r, str>> {
     fn request_body(gen: &mut OpenApiGenerator) -> Result {
-        fn_request_body!(gen, str, "application/octet-stream")
-    }
-}
-impl<'r> OpenApiFromData<'r> for rocket::data::Capped<&'r str> {
-    fn request_body(gen: &mut OpenApiGenerator) -> Result {
-        fn_request_body!(gen, str, "application/octet-stream")
-    }
-}
-// See: https://github.com/GREsau/schemars/issues/103
-impl<'r> OpenApiFromData<'r> for rocket::data::Capped<&'r rocket::http::RawStr> {
-    fn request_body(gen: &mut OpenApiGenerator) -> Result {
-        <&'r rocket::http::RawStr>::request_body(gen)
-    }
-}
-impl<'r> OpenApiFromData<'r> for rocket::data::Capped<&'r [u8]> {
-    fn request_body(gen: &mut OpenApiGenerator) -> Result {
-        Vec::<u8>::request_body(gen)
-    }
-}
-impl<'r> OpenApiFromData<'r> for rocket::data::Capped<String> {
-    fn request_body(gen: &mut OpenApiGenerator) -> Result {
-        String::request_body(gen)
-    }
-}
-impl<'r> OpenApiFromData<'r> for rocket::data::Capped<Vec<u8>> {
-    fn request_body(gen: &mut OpenApiGenerator) -> Result {
-        Vec::<u8>::request_body(gen)
+        <Cow<'r, str>>::request_body(gen)
     }
 }
 
@@ -128,16 +85,59 @@ impl<'r> OpenApiFromData<'r> for &'r rocket::http::RawStr {
     }
 }
 
-impl<'r> OpenApiFromData<'r> for Data<'r> {
+impl<'r> OpenApiFromData<'r> for String {
     fn request_body(gen: &mut OpenApiGenerator) -> Result {
         fn_request_body!(gen, String, "application/octet-stream")
+    }
+}
+
+impl<'r> OpenApiFromData<'r> for Vec<u8> {
+    fn request_body(gen: &mut OpenApiGenerator) -> Result {
+        fn_request_body!(gen, Vec<u8>, "application/octet-stream")
+    }
+}
+
+impl<'r> OpenApiFromData<'r> for rocket::data::Capped<&'r str> {
+    fn request_body(gen: &mut OpenApiGenerator) -> Result {
+        <&'r str>::request_body(gen)
+    }
+}
+
+// See: https://github.com/GREsau/schemars/issues/103
+impl<'r> OpenApiFromData<'r> for rocket::data::Capped<&'r rocket::http::RawStr> {
+    fn request_body(gen: &mut OpenApiGenerator) -> Result {
+        <&'r rocket::http::RawStr>::request_body(gen)
+    }
+}
+
+impl<'r> OpenApiFromData<'r> for rocket::data::Capped<&'r [u8]> {
+    fn request_body(gen: &mut OpenApiGenerator) -> Result {
+        <&'r [u8]>::request_body(gen)
+    }
+}
+
+impl<'r> OpenApiFromData<'r> for rocket::data::Capped<String> {
+    fn request_body(gen: &mut OpenApiGenerator) -> Result {
+        String::request_body(gen)
+    }
+}
+
+impl<'r> OpenApiFromData<'r> for rocket::data::Capped<Vec<u8>> {
+    fn request_body(gen: &mut OpenApiGenerator) -> Result {
+        Vec::<u8>::request_body(gen)
+    }
+}
+
+impl<'r> OpenApiFromData<'r> for Data<'r> {
+    fn request_body(gen: &mut OpenApiGenerator) -> Result {
+        Vec::<u8>::request_body(gen)
     }
 }
 
 // `OpenApiFromForm` is correct, not a mistake, as Rocket requires `FromForm`.
 impl<'r, T: JsonSchema + super::OpenApiFromForm<'r>> OpenApiFromData<'r> for rocket::form::Form<T> {
     fn request_body(gen: &mut OpenApiGenerator) -> Result {
-        fn_request_body!(gen, T, "application/octet-stream")
+        fn_request_body!(gen, T, "multipart/form-data")
     }
 }
 
@@ -153,5 +153,20 @@ impl<'r, T: JsonSchema + Deserialize<'r>> OpenApiFromData<'r>
 {
     fn request_body(gen: &mut OpenApiGenerator) -> Result {
         fn_request_body!(gen, T, "application/msgpack")
+    }
+}
+
+impl<'r, T: OpenApiFromData<'r> + 'r> OpenApiFromData<'r> for StdResult<T, T::Error> {
+    fn request_body(gen: &mut OpenApiGenerator) -> Result {
+        T::request_body(gen)
+    }
+}
+
+impl<'r, T: OpenApiFromData<'r>> OpenApiFromData<'r> for Option<T> {
+    fn request_body(gen: &mut OpenApiGenerator) -> Result {
+        Ok(RequestBody {
+            required: false,
+            ..T::request_body(gen)?
+        })
     }
 }