+++ rocket-okapi/src/request/from_segments_impls.rs
@@ -1,3 +1,6 @@
+//! Implement `OpenApiFromSegments` for everything that implements `FromSegments`
+//! https://docs.rs/rocket/0.5.0-rc.2/rocket/request/trait.FromSegments.html#foreign-impls
+
 use super::OpenApiFromSegments;
 use crate::gen::OpenApiGenerator;
 use okapi::openapi3::{Object, Parameter, ParameterValue};